refactor : 중복되는 코드 제거 및 컨벤션 수정
