+++ src/main/java/chess/domain/piece/pathStrategy/PathStrategy.java
@@ -1,11 +1,16 @@
 package chess.domain.piece.pathStrategy;
 
-import chess.domain.board.Position;
-
 import java.util.List;
 
-public interface PathStrategy {
-    void validateDistance(Position sourcePosition, Position targetPosition);
+import chess.domain.board.Position;
 
-    List<Position> createPath(Position sourcePosition, Position targetPosition);
+public abstract class PathStrategy {
+	public List<Position> createPaths(Position sourcePosition, Position targetPosition) {
+		validateDistance(sourcePosition, targetPosition);
+		return findPaths(sourcePosition, targetPosition);
+	}
+
+	abstract protected void validateDistance(Position sourcePosition, Position targetPosition);
+
+	abstract protected List<Position> findPaths(Position sourcePosition, Position targetPosition);
 }
